+++ src/fields/SelectField.vue
@@ -12,6 +12,7 @@ import {
 import { useRouter } from '@/router'
 import { onClickOutside } from '@vueuse/core'
 import { ICON_NAMES } from '@/enums'
+import { useI18n } from 'vue-i18n'
 
 type SCHEMES = 'primary'
 type SIZES = 'medium' | 'default'
@@ -31,6 +32,7 @@ const props = withDefaults(
     placeholder?: string
     errorMessage?: string
     isErrorMessageShown?: boolean
+    isNeedAllOption?: boolean
   }>(),
   {
     scheme: 'primary',
@@ -41,6 +43,7 @@ const props = withDefaults(
     placeholder: ' ',
     errorMessage: '',
     isErrorMessageShown: true,
+    isNeedAllOption: false,
   },
 )
 
@@ -49,6 +52,7 @@ const emit = defineEmits<{
 }>()
 
 const attrs = useAttrs()
+const { t } = useI18n({ useScope: 'global' })
 
 const selectElement = ref<HTMLDivElement>()
 
@@ -71,8 +75,15 @@ const isReadonly = computed(() =>
 
 const isLabelActive = computed(() => isDropdownOpen.value || !!props.modelValue)
 
+const options = computed(() => [
+  ...(props.isNeedAllOption
+    ? [{ value: '', label: t('select-field.all-lbl') }]
+    : []),
+  ...props.valueOptions,
+])
+
 const selectedOption = computed(() =>
-  props.valueOptions.find(i => i.value === props.modelValue),
+  options.value.find(i => i.value === props.modelValue),
 )
 
 const selectFieldClasses = computed(() => ({
@@ -169,7 +180,7 @@ watch(
             />
           </template>
           <template v-else>
-            <template v-if="modelValue">
+            <template v-if="modelValue || isNeedAllOption">
               <div class="select-field__select-head-value">
                 <icon
                   v-if="selectedOption?.icon"
@@ -222,7 +233,7 @@ watch(
                 }"
               />
             </template>
-            <template v-else-if="valueOptions.length">
+            <template v-else-if="options.length">
               <button
                 :class="[
                   'select-field__select-dropdown-item',
@@ -232,7 +243,7 @@ watch(
                   },
                 ]"
                 type="button"
-                v-for="(option, idx) in valueOptions"
+                v-for="(option, idx) in options"
                 :key="`[${idx}] ${option.value}`"
                 @click="select(option.value)"
               >