+++ src/pages/MyOrders/CreatedOrders.vue
@@ -14,6 +14,7 @@
         v-model="networkId"
         scheme="primary"
         size="medium"
+        is-need-all-option
         :disabled="isLoading"
         :value-options="chains"
       />
@@ -109,10 +110,7 @@ const isListEmpty = ref(true)
 const isLoadFailed = ref(false)
 const isLoading = ref(false)
 const isCreateOrderModalShown = ref(false)
-const networkId = ref(
-  chainStore.selectedChain?.chain_params.chain_id ??
-    chainStore.chains[0].chain_params.chain_id,
-)
+const networkId = ref<string | number>('')
 
 watch(isButtonFixed, val => {
   emit('is-button-fixed', val)